cloud_functions/audio_generator/requirements.txt
functions-framework==3.5.0
google-cloud-texttospeech==2.14.0
google-cloud-storage==2.14.0
